gap fitter is not working, moving on to different approach
Former-commit-id: b97c99b67d73070e9584d503ab972212e1fa5c15

+++ src/Game/Data.hs
@@ -48,7 +48,7 @@ data Action = NullAction
             | Idle { pos :: (Float, Float) }
             | Loaf { time :: Int }
             | MoveTo { dest  :: (Float, Float)
-                     , speed :: Int
+                     , speed :: Float
                      } deriving (Show, Eq)
 
 data Unit = Unit { unittexs :: [GL.TextureObject]
